Bug 19026 - Types with same name and different size considered equivalent
This is a second attempt at fixing this bug.  The previous attempt was
on the good path, but a thinko got in the way, oops.  We are not adding
the new binary of that bug report (which got re-opened) to the test
suite because it's too big and it takes too much time and memory to
abidw to analyze it.

	* src/abg-ir.cc (type_base::get_canonical_type_for): Really
	compare the size of the type to be canonicalized against the size
	of the *current* potential canonical type of the same name.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>