[PERF] Speedup comparing declaration-only class_decls
The structural comparison of declaration-only class_decl has show up
high on performance profiles while running abidw --abidiff on the
r300_dri.so binary.

The class_decl::get_is_declaration_only() function seems to be used
too many times.  The copying of the qualified name of the class_decl
showed up too.  Also, it looks like that the != operator on string is
slightly slower than than the == operator; OK, this is while looking
at a binary without optimizations, but still.

So the patch addresses the specific issues above.

	* src/abg-ir.cc (equals): In the overload for class_decl, avoid
	calling class_decl::get_is_declaration_only() several times.
	Avoid copying the qualified name of the class_decl.  Also, use the
	== operator to compare strings, rather than the != one.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>