Support two different variables having the same underlying symbol
Sometimes, two different variables A and B can have the same underlying
symbol, when for instance A is a reference that is initialized with
the address of B.

Until now, that was not supported by libabigail's ABI corpus.  Only
one (either A or B) variable is kept.  This is because the ID of the
variable, as returned by abigail::var_decl::get_id() is made of the
symbol name of the variable only.

This patch fixes the issue by including the name of the variable in
the ID.

	* src/abg-ir.cc (var_decl::get_id()): Include the name of the
	variable in the ID.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>